pyproject.toml
[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "astrobasecloud"
authors = [{name = "Astrobase Inc.", email = "anthony@astrobase.cloud"}]
dynamic = ["version", "description"]
readme = "README.md"
requires-python = ">=3.7.5"
dependencies = [
    "fastapi>=0.74.0",
    "typer>=0.4.0",
    "gunicorn>=20.1.0",
    "uvicorn>=0.17.0",
    "PyYaml>=6.0",
    "google-cloud-container>=2.10.3",
    "google-cloud-service-usage>=1.2.1",
    "boto3>=1.17.22",
    "azure-identity>=1.7.1",
    "azure-mgmt-resource>=20.1.0",
    "azure-mgmt-containerservice>=17.0.0",
    "msrestazure>=0.6.4"
]

[project.urls]
Home = "https://www.github.com/astrobase/astrobase"
Documentation = "https://astrobase.cloud"

[project.scripts]
astrobase = "astrobasecloud.cli.main:app"

[project.optional-dependencies]
test = [
    "pytest>=6.2.5",
    "mypy>=0.910",
    "coverage>=6.2",
    "pytest-cov>=3.0.0",
    "requests-mock>=1.9.3",
    "types-PyYAML>=6.0.3",
    "types-requests>=2.27.8"
]
doc = [
    "mkdocs>=1.1.2,<2.0.0",
    "mkdocs-material>=8.1.4,<9.0.0",
    "mdx-include>=1.4.1,<2.0.0",
]
dev = [
    "flake8>=4.0.1",
    "black>=21.12b0",
    "isort>=5.10.1",
    "autoflake>=1.4",
    "flake8-docstrings>=1.6.0"
]

[tool.isort]
profile = "black"

